backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "claude-agent-sdk>=0.1.6",
    "fastapi>=0.121.2",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
    "x402>=0.2.1",
]
